+++ docs/GettingStarted/1. Introduction.mdx
@@ -1,7 +1,18 @@
 import { Meta, Story} from '@storybook/addon-docs';
-import * as HeroStory from "../src/Hero.stories";
-import Hero from "../src/Hero";
+import * as HeroStory from "../Hero/Hero.stories";
+import Hero from "../Hero/Hero.stories";
 
 <Meta title="Getting Started/Introduction"/>
 
-<Hero />
+<Story of={HeroStory.Default} />
+
+
+
+`react-hexgrid` helps you create tessellations of regular hexagons without the headache of wrangling with SVGs.
+
+It uses `<g>` and `<polygon>` tags to make regular hexagons - no fancy CSS tricks!  Your hexagons will be easy to read, style, and maintain.
+
+There are also algorithms for generating common hexagon patterns.
+
+Boasting 0 dependencies, full TypeScript support, Storybook visual tests, and a commitment to no breaking changes, any react-hexgrid code you write 
+will last a long, long time.
