+++ docs/Hero/main.tsx
@@ -1,3 +1,4 @@
+import React from "react"
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import Hero from './Hero.tsx'
