--- docs/test.mdx
@@ -1,6 +1,0 @@
-import { Meta, Story } from '@storybook/addon-docs';
-import * as HeroStory from "../src/Hero.stories";
-
-<Meta title="Getting Started/Introductioasdf"/>
-
-<Story of={HeroStory.Default} />
